Mouse shooting

+++ src/com/gibirus/entities/Shoot.java
@@ -9,12 +9,12 @@ import com.gibirus.world.Camera;
 
 public class Shoot  extends Entity{
 
-	private int dx;
-	private int dy;
+	private double dx;
+	private double dy;
 	private double spd = 4;
 	
 	private int life = 40, curLife = 0;
-	public Shoot(int x, int y, int width, int height, BufferedImage sprite, int dx, int dy) {
+	public Shoot(int x, int y, int width, int height, BufferedImage sprite, double dx, double dy) {
 		super(x, y, width, height, sprite);
 		this.dx = dx;
 		this.dy = dy;
